+++ src/cq_warehouse.egg-info/PKG-INFO
@@ -5,9 +5,7 @@ Summary: A cadquery parametric part collection
 Home-page: https://github.com/gumyr/cq_warehouse
 Author: Gumyr
 Author-email: gumyr9@gmail.com
-License: UNKNOWN
 Project-URL: Bug Tracker, https://github.com/gumyr/cq_warehouse/issues
-Platform: UNKNOWN
 Classifier: Programming Language :: Python :: 3
 Classifier: License :: OSI Approved :: Apache-2.0 License
 Classifier: Operating System :: OS Independent
@@ -37,4 +35,3 @@ To install **cq_warehouse** from github:
 ```
 python3 -m pip install git+https://github.com/gumyr/cq_warehouse.git#egg=cq_warehouse
 ```
-
